test to illustrate problem
The first step in fixing the problem is to write a unit test to demonstrate it. This test fixture is constructed to have an __all__ target in it making the unit test rather simple -- test_issue_16().

The complexity is really in the setUpClass(). Here it overrides critical parts of dawgie.db.shelf that normally speak over the network via twisted. The unit test does not require the communication but does require the standard processing. Hence, setattr() is used to do the overrides:

mock_acquire and mock_release override dawgie.db.shelf.Connector versions because there can only be one user of the database -- pytest. Therefore, the locks can simply be ignored.

mock_do() overrides dawgie.db.shelf.Connector.__do() skipping the pickling sending and receiving. Instead it calls dawgie.db.shelf.Worker.do() directly and the response is passed back via a test global variable.

mock_send() overrides dawgie.db.shelf.Worker._send() to skip pickling and sending. Instead it writes the respose to a test global variable.