unit test
dawgie.tool.compliant is the unit test for this problem. Seems that not all SV_REF and V_REF are created equal and this will check a system that they all comply with the expectation. Given that the tests have a toy AE in them, this deployment check also becomes a unit test.