demonstration of fault
A unit test that uses the toy test/ae to demonstrate the detection of wrong types in *_REF. All of hte ALG_REF work without problem. It is the newer and much less used SV_REF and V_REF. Currently it is finding the errors in test/ae as expected.