Issue 18: ref verification (#19)
* the feat should always be a string

While the documentation stated object, dawgie expects it to be a string because dawgie.StateVector uses a string (feature name) for the key. Secondary, the dawgie.StateVector key is used to form a canonical name with the feat being the last part.

* unit test

dawgie.tool.compliant is the unit test for this problem. Seems that not all SV_REF and V_REF are created equal and this will check a system that they all comply with the expectation. Given that the tests have a toy AE in them, this deployment check also becomes a unit test.

* demonstration of fault

A unit test (Test/test_08.py) that uses the toy test/ae to demonstrate the detection of wrong types in *_REF. All of the ALG_REF work without problem. It is the newer and much less used SV_REF and V_REF. Currently Test/test_08.py is finding the errors in test/ae as expected.

* update toy ae in Test

ae.feedback.bot updated so that all SV_REF and V_REF follow the dawgie documentation (doctext in dawgie/__init__.py). They now set all items to dawgie.StateVector instead of strings. The result was to cause deeper parts of dawgie to fail then be corrected.

* clean up dawgie

dawgie.pl.dag -- all ref.item are now treated as a dawgie.StateVector
dawgie.util -- all ref.item and vref.item are treated as dawgie.StateVector

* final validation

With all of the changes above completed and tested via .ci/check*.sh passing, ran the scripts .ci/deploy_0[1-6].sh inclusive. Fixed a trivial poor use of git checkout to git reset in .ci/deploy_01.sh that is non sequitur. Proceeded to run .ci/exercise_01.sh after all of the deploy scripts completed. The exercise ran to completion verifying that dawgie.db is sufficient as is and that all other changes work in concert correctly.