issue 7: FE updates (#10)
* Updating FE with Build v19

Changed typeahead engine to use fuzzy matches on the search page.

Made State Vector search use filtering the same as Target Name search.

* Used the `.ci/check_04.sh 1` script to update the LICENSE information in the html files.

* change default filters

The Admin filter is working in tests so should be sufficient. The Dev and User are set to be the same by default. Still testing with those so they may update again in the future.

* Re-arragning filter buttons.

Fixing a bug in filters for exact match.

* reset user and dev default values now that filters are set correctly

* Adding a cookie for selected filter state

Adding a way to brand the frontend with javascripts/siteconfig.js

* Removing un-used images