start of deploying via PyPI
An instiall script to build the a deployment. Right now the script builds this directory rather than collecting a released version from github.com. It will come with time, but here the mechanics of how to build the artifact for PyPI and deliver it is complete. The script is left in safe mode where it will deploy to test.pypi instead of pypi.

Updated the setup.py to work correctly with PyPI. It does a description (poorly) but does one. It has a license and describes it nicely on the website. All inall, it looks like the setup.py is sufficient.