Issue 27: expand compliant checks (#28)
* add rule 9: no state vectors

First, add the rule_09()  to tools/compliant.py. It simply runs through the factories and makes sure that all units return an non-zero length array.

Second, add a negative test to Test/test_08.py along with a bad AE (bae). Since there is a positive test already to verify Test/ae will work, the negative test will verify that rule 9 is coded correctly.

Lastly, using the test with the bae correct the code and test until it all works.

* non sequitour change

Updated dawgie.tools.dag to have a verbose switch. That is when it was discovered that state_vectors() must not be zero length to get added to the dag. Still, these changes would be nice to expand upon in the future so keeping them here.