Issue 29: filter defaults (#34)
* non-sequitur changes

Added some logging to help understand why navel gazing is taking so long. They were needed to debug the private pipeline to the point were this issue could be tested.

* Improper handling of object or string

dawgie.fe.app.search_filter_{a,d,u}() were not uniformly handling the type of the object. Some parts treated it as a dictionary others a string. Made it a dictionary consistently until it is turned to JSON at the return.