Issue 47: backlog (#48)
* add configurable TCP/IP backlog

As the worker size scales up, the backlog needs to increase with it. Since it should be directly related to the number of workers, called it worker_backlog.

* update all worker interfaces

There are many connections between workers and the pipeline server. Update them all to use a big enough backlog to support the scale of workers being used.