add capabilities
dawgie.MOMENT added "boot" which will trigger something when the pipeline boots and only boots (not reload from git). This is an odd request but since there is a need, added it.

dawgie.schedule() was updated to properly handle the new "boot" in dawgie.MOMENT. The check was modified to be more adaptable when dawgie.MOMENT changes again in the future.

dawgie.tools.compliant was updated to check that events() returns valid dawgie.MOMENTS since nothing requires the user use dawgie.schedule().