adjust factories (#65)
dawgie.fe now uses the default display rather than hard coded to dawgie.de.html. This allows projects to build their own visitors since this task is generally tedious and user specific. The current dawgie.de.html is the bare minimum.

dawgie.de.factory() now checks the input for None (default) and uses the module specified in dawgie.context.display. For backward compatibility, if no "." are detected in the module name, it prepends dawgie.de to it. Override with care!