reworked deferred requests
Rendering a state vector was already a deferred task. Often these renderings take a significant amount of time so they are allowed to happen in a separate thread so that the GUI is still responsive. However, it was done in an ad-hoc way that allowed for only it to be done as a deferred. Reworked the concept to allow other tasks to become deferred as well.

Added dawgie.fe.svrender to handler the work of deferring the task to another thread in this module. It now isolates that the deferred in in another thread in this module and how it closes out the work in this module as well.

Updated dawgie.fe.app to import the new module and load the dawgie.fe.DynamicContent with the new methods and objects.

Cleaneup up redundant code from dawgie.de.