issue 80: fix auto merge (#81)
* pass gpghome

Passing the ENV was not enough in this situtation. Instead, need to add the ability to pass the running gpg home to the spawned compliancy check.

* reworked deferred requests

Rendering a state vector was already a deferred task. Often these renderings take a significant amount of time so they are allowed to happen in a separate thread so that the GUI is still responsive. However, it was done in an ad-hoc way that allowed for only it to be done as a deferred. Reworked the concept to allow other tasks to become deferred as well.

Added dawgie.fe.svrender to handler the work of deferring the task to another thread in this module. It now isolates that the deferred in in another thread in this module and how it closes out the work in this module as well.

Updated dawgie.fe.app to import the new module and load the dawgie.fe.DynamicContent with the new methods and objects.

Cleanup up redundant code from dawgie.de.

* refactor code

After pulling out the subprocess.spawn() from dawgie.tools.submit.auto_merge() had to rework the code from the app hook down to become async. This should also free up some part of the GUI response when a submit happens. Not sure though because it may be happening during the reload rather than this portion.

Move the dawgie.pl.start.submit() to dawgie.fe.submit.Process along with the break up of dawgie.tools.submit.auto_merge() in support as well. It is less obvious but is much more async friendly.