Issue 76: support https (#79)
* required for twisted to do https

* add hook for cert

Seems that all that is needed is the SSL PEM file as it contains both the cert and the key. Twisted uses this single file to initiate an https server.

* clean up in preparation

Found two mains in dawgie.fe. Removed what appears to be the older one and it looks fine in comparison to the actual main.

* add https hooks

When starting the gui, check to see if a pem file exists. If it does and is a real file, then try to use it as the cert. While this matches the requirements in the twisted documenation it still needs to be tested somehow.

* update rules

twisted.internet.reactor.listenSSL is generated at run time.

* runs standalone with https

The PEM cert and key must be in a single file. Once that was done, it worked.