complete email alerts
Updated dawgie.tools.submit to use dawgie.context for the email list. If the list is not defined, then just log it. In fact, the message will always be in the log file for those not on the mailing lists.

Updated dawgie.fe.submit to reflect changes in dawgie.tools.submit.