Issue 84: remove hardcoded email alerts (#89)
* add email alert destination

The first part of the problem is to add a configuration or context driven destination for alerts via email. This was done by adding dawgie.context.email_alerts_to that follows the usual paradigm.

* complete email alerts

Updated dawgie.tools.submit to use dawgie.context for the email list. If the list is not defined, then just log it. In fact, the message will always be in the log file for those not on the mailing lists.

Updated dawgie.fe.submit to reflect changes in dawgie.tools.submit.