Issue 90: excise SDP (#92)
* change dawgie.pl, dawgie.tools, dawgie.fe, and unit testing

Updated farm, start, and state such that all SDP and sdp became FSM and fsm respectively.

* update db modules as well

The shelf and util modules were different then the previeous alterations where SDP was used to rever to an FSM. In this case, sdp was hardcoded as the db_name. Hence, all reference where changed to dawgie.context.db_name.

* update pylint RC

Many of the FSM functions are added at run time by the transistions package. Had to change sdp to fsm in the pylint RC to account for this.