Issue 91: refactor dawgie.pl.start (#96)
* rename

* clean up old start

Changed the indentation and organized the code now that the code is not in its own module. Moved the two support functions _main() and _merge() to __init__.py and then impoted them locally to __main__.py.

* change old dynamic generated references

pylint has a hard time with the fsm because most of it is added at run time by the transitions module. Updated from pl.start to context and they should now all work out cleanly.

* clean up perphial modules

Now that dawgie.pl is in good condition. Need to use pylint to find all of the reference to the old dawgie.pl.start.fsm and replace them with dawgie.context.fsm.

* need to pass the args

* rearrage startup some more

* need an import

* fix typo