extend rule 01
The factory methods that dawgie depends on were not being fully checked. The number of arguments was being checked but that let errors creep through with defaults and types. Some parts of the system require default values for the factory arguments. Therefore, extended the rule 01 testing to be parameter count first, then default values followed by types.