export versions
Update setup.py to output the list of modules to dawgie/fe/requirements.txt with just the module names and nothing else. The app can use that file to look up the current versions.

Update dawgie.fe.app to add a new endpoint (URL) that outputs a JSON object with a list of the current versions.