issue 114: improve version information (#115)
* export versions

Update setup.py to output the list of modules to dawgie/fe/requirements.txt with just the module names and nothing else. The app can use that file to look up the current versions.

Update dawgie.fe.app to add a new endpoint (URL) that outputs a JSON object with a list of the current versions.

* change tactic

Moved from using importlib to pip. The distribution names are not the same as the package names. Since the requirement file is being used to find the versions, then just use pip all the way instead of guessing package names from it.