Issue 111: redo LogDeferredException (#118)
* refactored dawgie.pl

LogDeferredException was changed to two classes: LogFailure and DeferWithLogOnError. The first does nothing but log the failure including a stack trace if it is an exception. The latter registers the callback and hooks in the previous log on an error. Removed LogDeferredException.

* fix references to dawgie.pl.LogDeferredException

Some modules, dawgie.schedule and dawgie.pl.worker.aws, use DeferWithLogOnError. They need it to create the deferred then jump to the given callback (cb) that takes no arguments.

The others use just the logging use LogFailure. .

* fix up main

With all of the other problems fixed, restructure __main__. Defined the correct class in the main itself since no other class should be using it. Fixed up all of the naming after that.