Issue 103: make aspects quack like dicts (#121)
* rearrange db util

To make space for the new aspect implementations. Created a package dawgie.db.util then moved the module dawgie.db.util to the package so that the existing code could remain unmodifiled.

* Update the aspect interface

Add the container methods the aspect should implement. They keep the container read-only.

* general implementaiton of container needs

Most of the work of the container can be done via the standardieze module that passes the buck at the critical moment. Particularly when doing the lazy loading of the data from the database.

* dawgie.Timeline as well

Just as dawgie.Aspect behaves like a write only dictionary so should timeline. Updated the comments and added the appropriate abstract functions.

* update implementations

With the framework in place, replace the current implementations for both post and shelf to implement the dictionary as documented. Since both dawgie.Aspect and dawgie.Timeline behave roughly the same, then implementation should work for both.

* fix dawgie.context.dumps()

Because of problems in a project included a unit test to verify that dumps() is working. Fixed dumps() when problem became apparent.