Issue 129: implement externalization of events (#131)
* externalize events

The front-end should publish the known event list and duration to next event. This is as simple as adding a new DynamicContent element that makes the endpoint.

* implement event externalization

Convert the list of periodic events (nodes that contain a list of moments) to a list of dictionaries that are the actor and delays. These are then converted to a list of objects in JSON by the front-end for it to display.

* add unit testing

Verify dawgie.pl.schedule.view_events() correctly converts the periodic list using a known set of events and momements when they should happen.