update new values
dawgie.Analysis, dawgie.Regress, and dawgie.Task are all updated to expect a tuple of name,isnew instead of just name. dawgie.pl.worker already handles the list as an object and simply pass it along requiring no changes.

dawgie.db.post and dawgie.db.shelf update the list and are updated as well. Instead of adding the list conditionally, they always add to the list and pass along if it is new or not.

dawgie.pl.schedule was updated to work with the tuple. It behaves as it always has or at least that is the intent but now allows for expansion in ability.