Issue 136: add promotion hooks (#138)
* alter scheduler to enable auto promotion

Added hooks to dawgie.pl.schedule to trigger the auto promotion. Insterted the call to the dawgie.pl.schedule.update() which will add items as required by results sent from dawgie.pl.worker. Added a conditional to dawgie.pl.schedule.next_batch() that will do promotion before scheduling new jobs so that the new jobs have the current best estimate to work with.

Added dawgie.pl.auto which contains the stub for Promote that keeps the current behavior but concentrates where changes need to be made for auto promotion.

* add unittest

While there is not much to test, add what can be done. The first and basically only interesting test is to see that __call__ is working as expected. Wrong ars are processed correctly as is more. Log message shows up correctly as well.

* rename it all

Unhappy with dawgie.pl.auto.Promote renamed it to dawgie.pl.promotion.Engine.

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>