mark this point
It seems this is mostly wrong but want to mark this point in case the new idea is wrong and need to go back. The changes will remove the constraint first and make it based on just he output and what it will require to run. If it has been run with those conditions previously then use just the already known output. Does not require the complicated constraints and inputs that this edition has partially impleemted.