promotion should find junctures
Worked consistent() a couple of times to finally arrive at this algorithm should select the most recent equivalent output that matches the current inputs. Left comments in place to highlight the algorithm. It bases its decision on the current version of the algorihm and its output over time. Using the input it would have now if it ran, it then uses the output runids to check for inputs that match what would be the lastest inputs.

The most interesting point, is that promotion should go and this should be done always for all items. However, shelf cannot currently support this level of interigation. However, it still should be scheduled for the next big releaes of dawgie to replace the current scheduler.