update testing
The bulk of the postgres database interface has been tested. Two function are ignored because postgres does not need to implement them: copy() view_locks(). It is the baseline for comparing the various supported databases to make them more or less plug and play.

The two remaining functions consistent() and promote() are needed for this specific ticket but the others show that the postgres interface is robust and stable.