logging seems to have changed
Somehow, maybe pytest, logging.basicConfig() is being called and so the handler is no longer being set. Changed to add the handler to the child used for testng. Since the main probgram has more control over logging.basicConfig() and issue 45 focused on the handlers behavior, this change so keep the spirit of the test while ignoring updates to external tools.