expand check_03,sh setup
Now that promotion can be checked for postgres, it is important to load and install postgres during check_03.sh. Since it uses the postgres:latest docker image to complete this work, the expansion of check_03.sh is direct. It first checks to see if the docker image is present then the docker network then it starts the container and creates the needed database. When complete, it stops the docker container.