Issue 139: implement promotion engine (#144)
* finer control over promotion

Added the switch dawgie.context.allow_promotion that can turn it on and off easily. There are conditions that if not followed will really foul up promotion. The belief is that EXCALIBUR might currently be existing in those conditions such that promotion will be bad for them.

* extend unittest

The unit testing PromotionEngine required extension to properly test the addition of dawgie.context.allow_promotion.

* extend promotion

dawgie.pl.promotion to use the current AE to find the dependencies.

dawgie.pl.context changed to support the new allow_promition.

updated test 12 to handle the allow_promotion flag with if the AE is set.

* skeletal promotion

* mark this point

It seems this is mostly wrong but want to mark this point in case the new idea is wrong and need to go back. The changes will remove the constraint first and make it based on just he output and what it will require to run. If it has been run with those conditions previously then use just the already known output. Does not require the complicated constraints and inputs that this edition has partially impleemted.

* promotion should find junctures

Worked consistent() a couple of times to finally arrive at this algorithm should select the most recent equivalent output that matches the current inputs. Left comments in place to highlight the algorithm. It bases its decision on the current version of the algorihm and its output over time. Using the input it would have now if it ran, it then uses the output runids to check for inputs that match what would be the lastest inputs.

The most interesting point, is that promotion should go and this should be done always for all items. However, shelf cannot currently support this level of interigation. However, it still should be scheduled for the next big releaes of dawgie to replace the current scheduler.

* promition adds entries

Now that dawgie.db.consistent() returns what should be a valid juncture, check then install the juncture. promotion() was updated to reflect a possible failure of the juncture insert will force the item to run.

All checks pass but unit testing is not done.

* update all files for license/copyright updates

This update is a pain and probably unnecessary but it is nice in that forces an update to containers, external dependencies, etc that can be easily forgotten. In updating the Dockerfile.cit moved to more modern tools that worked with a newer Ubuntu 20.04 in Dockerifle.os. Many other changes took place after these, but they were the start.

* .ci/check_03.sh updated for new tests

Test/test_13.py was added to do the bulk of the promotion testing within the databases themselves. Right now, shelf is not able to support promotion so postgres is the only database being tested. Test/test_13.py will also work well to have consistent behavior among all supported databases.

To support Test/test_13.py, .ci/check_03.sh was extended to download postgres image, start it, and prepare it (createdb) then clean up the container when finished.

* logging seems to have changed

Somehow, maybe pytest, logging.basicConfig() is being called and so the handler is no longer being set. Changed to add the handler to the child used for testng. Since the main probgram has more control over logging.basicConfig() and issue 45 focused on the handlers behavior, this change so keep the spirit of the test while ignoring updates to external tools.

* expand check_03,sh setup

Now that promotion can be checked for postgres, it is important to load and install postgres during check_03.sh. Since it uses the postgres:latest docker image to complete this work, the expansion of check_03.sh is direct. It first checks to see if the docker image is present then the docker network then it starts the container and creates the needed database. When complete, it stops the docker container.

* apparently bridge is not good so default to citnet

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>