Issue 160: update base OS and all tests (#161)
* docker images build

Updating to Ubuntu 22.04 required newer versions of many of the tests like pylint as well. These changes required a bit tinkering to get all tools at a level where they are happy to work together.

* pylint improvements

The bulk of the upgrade goes to pylint. It has improved its checking depth and breadth. Code updates were required to clear the new and improved checks while some inline comments were more appropriate to the design of DAWGIE.

* remove some duplication

One of the newer checks is for code duplication. Other than the `__main__` block, duplicate blocks were removed and code cleaned up. Used an inline comment for the main blocks because they are naturally going be the same code over and over. The code is short and concise and really does not lend itself well to be being removed.