testing and db interface fixed
Seems there where some hidden problems with the db but not sure what caused the change to make them break suddnely.

Update the testdata.py because one of the string formats whas changed but forgot the f at the front of the string. That fixed all of the strange reductions in keys and such.

Update the post.py because either psycopg and/or psycopg2 both changed and what is returned is different than what it used to be. Anyway, all of test_13 is passing again.