issue 176: find actual cause (#177)
* fix two bugs

.ci/Dockerfile.py needs to use latest pydot that keeps up pyparsing changes

Test/test_10.py fixes a hidden dependency that test 2 ran already and set dawgie.context.fsm.

* cleaned up some testing

Seems with latest ubuntu and dawgie dependencies, some tests needed cleaned up after updating pydot.

Wrote a report for the compliance testing (check 5) too.

* update to latest psycopg

While it appears not to be necessary, updated to latest psycopg because psycopg2 seems to be at the end of it development. It did not fix the underlying problem with test 13 not did upgrading to latest postgres. Still, the changes are in a positive direction so keeping them.

* testing and db interface fixed

Seems there where some hidden problems with the db but not sure what caused the change to make them break suddnely.

Update the testdata.py because one of the string formats whas changed but forgot the f at the front of the string. That fixed all of the strange reductions in keys and such.

Update the post.py because either psycopg and/or psycopg2 both changed and what is returned is different than what it used to be. Anyway, all of test_13 is passing again.

* found the root cause (finally)

Can transition to Archiving from either Updating or Running. If from Updating, then the database is close. If from Running the database is open. Since Archiving needs the database open for the dawgie.tools.trace to run but closed for dawgie.db.archive(), Archiving needed to manage the open state of the database. Upgraded dawgie.db.reopen() to hint at its state when reopen() is called. This allows Archiving to set the database open state appropriately upon completion. This too was tricky because the trigger to exit Archving is called in _archive_done() not in _archive(). Therefore resetting the database open state has to be done in _archive_done() to avoid a race condition.