repair dawgie.db.post.trace() (#182)
The original code simply found the max for each of the different version values (design, impl, bugfix). The function needs to compute the max in a cascading for such that bugfix.max() is constrained by impl.max() which is then constrained by design.max() so that 1.1.7, 1.9.3 and 2.4.1 returns 2.4.1 and not 2.9.7. The latter is what the prefix code was doing and now the former is correctly returned with these changes.