Issue 183: extend interface (#184)
* add two helper functions to the API

It should be simple to add new targets. It is simple to use an Analyzer to compute all of the desireable targets. It is not easy to have that target added to the database without doing some dawgie gymnastics. Added dawgie.db.add() to remove these gymnastics from the AE.

The idea of sub-targets has come up. In the view of the AE some targets may have sub processing channels that are better shown with sub-target names like A(b) where A is the primary target and b is the sub-target. In this sense, algorihm foo() now works on sub targets A(b) and A(c) independently and in parallel rather than sequentially.

* add two helper functions to the API

It should be simple to add new targets. It is simple to use an Analyzer to compute all of the desireable targets. It is not easy to have that target added to the database without doing some dawgie gymnastics. Added dawgie.db.add() to remove these gymnastics from the AE.

The idea of sub-targets has come up. In the view of the AE some targets may have sub processing channels that are better shown with sub-target names like A(b) where A is the primary target and b is the sub-target. In this sense, algorihm foo() now works on sub targets A(b) and A(c) independently and in parallel rather than sequentially.

* update testing

The change in dawgie.db requires testing to show that any implementation for it works as expected.

* implementation of dawgie.db.add()

Post dawgie.db.post and dawgie.db.shelf required updating. Both have implementations and the postgres has been tested.