retarget implementation for postgres
Did the simplest/hardest all at once. Create duplicate target entries for upstream items by simply walking the prime table. It is a little lazy and it does the latest rid for a task and algorithm. If state vectors and values are well behaved as in defined at algorithm instantiation like they are supposed to be then this shortcut is sufficient.

Leaving shelf for when it is made to really behave like postgres.