renamed redirect to retarget (#192)
* renamed redirect to retarget

Fixed #190 as well since it was in the same location and easy.

Renamed the functions and checked that all connections still work. However, updating the past items still needs to be done.

* retarget implementation for postgres

Did the simplest/hardest all at once. Create duplicate target entries for upstream items by simply walking the prime table. It is a little lazy and it does the latest rid for a task and algorithm. If state vectors and values are well behaved as in defined at algorithm instantiation like they are supposed to be then this shortcut is sufficient.

Leaving shelf for when it is made to really behave like postgres.

* informational updates

Cleaned up error and log message. Changed the look of retarget() names.