Issue 189: redo shelf to shelve (#193)
* rewriting shelf

Changed a bunch of names (not all) from shelf to shelve while doing the rewrite just to make it match. It makes it much nicer in understanding the mapping of databases.

Cleaned up post2shelve to use the new shelve tables that account for versions.

Still had a bit of shelf laying about that testing found. Cleaned it up and changed some of the test harness to fit the new shelve.

The bulk of the changes came when the unit tests started running. It was here that most of the type changes and the whole concept of the db changes worked itself out.