Issue 194: updates to scheduler and testing of it (#205)
* first steps in fix

Improved documentation so the next time this code is looked at or reviewed,
it can go faster. It works out since what it does is no longer in my head and
this documentation represents what I found difficult to decipher.

* make all test runnable independently

Having to wait for all the tests to run was slowing down development. Wrapped it into this branch because this work is required sooner rather than later.

* fixed previous test that caused the problem

Turns out that promotion engine tests set something that caused this test to fail. Implemented a tearDownClass() to undo what it set. Fixed the tests when they all run.