add a snapshot capability
added a snapshot module to the dawgie.pl to build a dictionary that represents that moment in time. It grabs internal state from dawgie.context, dawgie.pl.state (fsm), dawgie.pl.farm, and dawgie.pl.schedule. It is really only helpful for debugging. Tested that items could be JSON serialized and the appear to be. Ready to move to ops for needed testing.

updated dawgie.pl.farm for more logging.

updated dawgie.fe.app so that can use curl to get the current state as a JSON object.