Issue 206: scheduler/farm/workers not getting things done (#207)
* allow chaining of check without network

The CI tests always failed if it could not update github. Changed that to fail on report content so can be chained nicely on the CLI without a network connection to github.com.

* add a snapshot capability

added a snapshot module to the dawgie.pl to build a dictionary that represents that moment in time. It grabs internal state from dawgie.context, dawgie.pl.state (fsm), dawgie.pl.farm, and dawgie.pl.schedule. It is really only helpful for debugging. Tested that items could be JSON serialized and the appear to be. Ready to move to ops for needed testing.

The updates include the implementation of the new endpoint /app/snapshot.