145: Fixes to promotion (#223)
* some simple fixes

There were basic problems when booting and promotion was active. Fixed the order that variables are set etc to boot and run with promotion. Fixed other typos in the promotion processing.

* need parents

The test for empty parents allow the code to progress, but it was not correct. Instead, dawgie.pl.dag needs to recreate the parents in the trimmed tree.

* updates to promotion

The original structure was to process the children of a node. Changed to add the children to the queue and process them directly. Those with a consistent data set are being promoted correctly.