Issue 226: set to unique fifo (#227)
* refactor util.py to util

Adding the FIFO to util is going to make the module cumbersome. Break it into its parts so that it can be simpler to understand what is related and what is not.

* implement and test a unique fifo

Instead of a set, it uses a set and list to back the concept. Made implementation pretty simple. Added a test to insure it works as desired.

* have todo use the unique fifo

Find all of the locations where todo is set to a set() and change it to dawgie.util.fifo.Unique(). The change was made in dawgie.pl.dag and dawigie.pl.schedule. Unit testing defines todo as a [] in some of tests. Where a set, it was changed to dawgie.util.fifo.Unique.