241: inter improvements (#242)
* correct args

Using list built rather than that supplied in args which means if given on the command line instead of stdin, it would not work. Cleaned up to use ARGS instead of ITEMS. Since postgres uses all the args, just pass them rather than pull some out as well

Added in a progress bar for both DB and DBS without option. The verbose switch is also used. Interring the data can take a while and it is good to understand just how big of block of data is being interred.