extend StateVector.view()
Added caller to StateVector.view(). This will cause a break with previous code because it adds a new paramter to the abstract call. It can easily be ignored by adding it _caller to all of the implementations. Pylint, or any static analyzwr really, should identify these locations and the trivial fix makes it an irritation not a break.