Issue 208: move to TLS (#246)
* update security to use TLS

The first step was to update dawgie.security to load TLS certificates and keys that define itself and guests that can have elevated privileges. Elevate privileges with respect to dawgie and AE it is running not the host that supports them.

Once the TLS were successfully loading, update the twisted sockets used to communicate with remote workers and the users (front-end or fe) to use the TLS certificates and keys. DAWGIE now uses the TLS client certificate to allow the logs, some DBs, and workers connect privately. Without the client certificate, requests to connect are refused.

DAWGIE still defaults to a fully open comminications (http and TCP/IP sockets). It must find 3 TLS certificates and keys for various layers of security. The security is built upon the previous layer so cannot have the 3rd without the first two:
1. sever certificate and key in a single PEM file defined with DAWGIE_SSL_PEM_FILE. It changes the front end from http to https.
2. certificate and key in single PEM file for dawgie internal communications defined with DAWGIE_SSL_PEM_MYNAME and DAWGIE_SSL_PEM_MYSELF. The first is the host name contained in the latter.
3. certificates only of the guests for visiting a running DAWGIE (front-end or fe) defined by certificates in a directory specified by DAWGIE_GUEST_PUBLIC_KEYS. To be discovered, the public keys must start with `dawgie.public.pem.` Because of my limited understanding of SSL and twisted, when guest certificates are found, a clone of the fe_port is made at fe_port + 5 (cfe for controlled front end) where client certificates are collected and used. Restricted endpoints are refused at fe_port.

* use PGP will be removed in a future release

Left the current PGP mechanisms in place. Not sure when it will be removed but probably next release.

* added a general sanctioned call

Added dawgie.security.sanctioned (endpoint, cert) to determine if a cert was given to all ow access to the endpoint. Returns True/False. Only a few endpoints are left of the generally and always sanctioned endpoints. 

is_sanctioned() and fetch_identity() are to be overridden via dawgie.context and thus the CLI args too. Allows an AE to adjust the access to the front-end endpoints to suit their needs.

If the overrides go awry, the default is to lock down the front-end until repaired.

* updated core dawige

Added caller to StateVector.view(). This will cause a break with previous code because it adds a new parameter to the abstract call. It can easily be ignored by adding it _caller to all of the implementations. Pylint, or any static analyzer really, should identify these locations and the trivial fix makes it an irritation not a break.

* tested the updates

Added unit testing for the loading of the TLS certificates and that the state is correctly detected.

Used `python -m dawgie.fe` to check that client certificates from google chrome do work correctly.

Updated the exercise scripts to use TLS and postgres and ensure it runs. This test generated two new tickets the most relevant of which is that CA support also needs to be added to this script. curl is very perticular and needs the CA certificate.

---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>